proba/requirements.txt
sqlite3
pillow
sv_ttk